requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
langchain==0.3.19
langchain_core==0.3.40
langchain_google_genai==2.0.11
python-dotenv==1.0.1
langchain-openai
mysql-connector-python
pandas
yfinance
pykrx
neo4j